+++ src/components/Parts/ComedienRoles/ComedienRoles.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./ComedienRoles.scss";
+
+export const ComedienRoles: React.FC = () => {
+    return (
+        <>
+            ComedienRoles
+        </>
+    );
+};