+++ src/components/Parts/PlayerName/PlayerName.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./PlayerName.scss";
+
+export const PlayerName: React.FC = () => {
+    return (
+        <>
+            PlayerName
+        </>
+    );
+};